[Feature] Always prompt the user before promoting a silver
By default the program will automatically promote pieces once they
reach the other side of the board.  That is almost always desirable.
However, it can be strategically advantageous to leave a silver
general unpromoted.  We can prevent the automatic promotion by
toggling 'Ask Before Promotion' in the Options menu, but that has the
tedious side-effect of requiring us to answer the prompt for every
possible promotion.

This patch implements a convenient compromise between the default
behavior and prompting for promotion: moving a silver general to the
opposite side of the board will always trigger the prompt, even if
'Ask Before Promotion' is disabled.  This lets us have automatic
promotion for all of the other pieces while also letting us control
whether or not we want to promote a silver.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>