[Feature] Support three pawn handicap games
The most severe handicap is for White to have nothing but his King,
i.e. the "naked king" handicap.  An unofficial step up from that is
for White to still have nothing but his King on board; but he begins
the game with three pawns in hand.  This lets White start by playing,
for example, a pawn drop to trick Black into losing his Bishop.

This patch implements the three pawn handicap style.  The video below
explains the nature of the handicap in more detail, three minutes in.

The actual implementation does nothing but add an SFEN notation for
the initial state of the game along with a new menu option.  The AI
tends to automatically go for that trick pawn drop, and so inherently
plays the handicap "correctly".

Signed-off-by: Eric James Michael Ritz <ericjmritz@yandex.com>

See-Also: https://www.youtube.com/watch?v=0w8_69flGjY&index=25&list=PL587865CAE59EB84A